+++ gui/.gitignore
@@ -74,3 +74,4 @@ Thumbs.db
 build/
 dist/
 venv/
+*.egg-info/
