+++ gui/systembridgegui/__init__.py
@@ -1,0 +1,1 @@
+"""System Bridge GUI: Init"""
